requirements.txt
google-cloud-firestore
requests
urllib3==1.26.6
